requirements.txt
grpcio==1.32.0
grpcio-tools==1.32.0
protobuf==3.13.0
Pillow==7.2.0
scikit-learn==0.22.2.post1
numpy==1.19.2
